Transform method completed in all states

+++ Graphics/Bullet.h
@@ -36,7 +36,7 @@ public:
 
 	// Methods
 	bool Move(int map[MAP_DIMENSION][MAP_DIMENSION]);
-	void Fire(int map[MAP_DIMENSION][MAP_DIMENSION], int securityMap[MAP_DIMENSION][MAP_DIMENSION]);
+	void Fire(int map[MAP_DIMENSION][MAP_DIMENSION], double securityMap[MAP_DIMENSION][MAP_DIMENSION]);
 	void Draw();
 };
